requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
Pillow
opencv-python
plyfile
tqdm
imageio
imageio-ffmpeg
